Add line length test to Markdown docs
